big bunch of changes
