+++ _inc/client/at-a-glance/style.scss
@@ -23,7 +23,7 @@
 .jp-at-a-glance__stats-inactive {
 	padding: rem( 16px );
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		display: flex;
 		flex-wrap: nowrap;
 		flex-direction: row;
@@ -32,12 +32,11 @@
 }
 
 .jp-at-a-glance__stats-inactive-icon {
-
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		display: none;
 	}
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		flex-basis: 10%;
 	}
 }
@@ -46,11 +45,11 @@
 	font-size: rem( 14px );
 	line-height: 1.5;
 
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		padding: 0 0 rem( 16px );
 	}
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		flex-basis: 50%;
 		padding: 0 rem( 16px );
 	}
@@ -59,7 +58,7 @@
 .jp-at-a-glance__stats-inactive-button {
 	text-align: left;
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		flex-basis: 40%;
 		text-align: right;
 	}
@@ -79,7 +78,7 @@
 .jp-at-a-glance__stats-bottom {
 	margin: rem( 32px ) 0 0;
 
-	@include breakpoint( "<480px" ) {
+	@include breakpoint( '<480px' ) {
 		box-shadow: 0 0 0 1px $light-gray-700;
 	}
 }
@@ -88,12 +87,12 @@
 	text-align: center;
 	border-bottom: 1px $light-gray-700 solid;
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		flex-wrap: nowrap;
 		display: flex;
 		flex-direction: row;
 	}
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		display: block;
 	}
 }
@@ -106,20 +105,20 @@
 }
 
 .jp-at-a-glance__stats-summary-today {
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		margin-top: rem( -1px );
 	}
 }
 
 .jp-at-a-glance__stats-summary-bestday {
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		margin: 0 rem( 1px );
 	}
 }
 
 .jp-at-a-glance__stats-summary-bestday,
 .jp-at-a-glance__stats-summary-alltime {
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		margin-top: rem( 1px );
 	}
 }
@@ -129,7 +128,7 @@
 	padding: rem( 16px );
 	box-shadow: 0 0 0 1px $light-gray-700;
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		max-width: 50%; // for IE 10 since it doesn't recognize flex-basis
 		display: flex;
 		flex-grow: 1;
@@ -146,20 +145,19 @@
 	padding: rem( 16px );
 	background-color: lighten( $gray, 35% );
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		display: flex;
 		flex-direction: row;
 		flex-wrap: nowrap;
 		align-items: center;
 	}
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		display: block;
 	}
 }
 
 .jp-at-a-glance__stats-cta-description {
-
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		flex-basis: 30%;
 	}
 }
@@ -175,13 +173,12 @@
 }
 
 .jp-at-a-glance__stats-cta-buttons {
-
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		text-align: right;
 		flex-basis: 70%;
 	}
 
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		text-align: center;
 
 		.dops-button {
@@ -211,7 +208,7 @@
 		outline: 0;
 	}
 
-	@include breakpoint( "<480px" ) {
+	@include breakpoint( '<480px' ) {
 		margin-left: 0;
 		margin-right: rem( 16px );
 	}
@@ -226,7 +223,7 @@
 	&:visited.is-current,
 	&:focus.is-current {
 		color: #23282d;
-		text-decoration:none;
+		text-decoration: none;
 	}
 }
 
@@ -239,7 +236,7 @@
 .jp-at-a-glance__item-grid {
 	display: flex;
 
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		display: block; // don't need flexbox on smaller screens
 	}
 }
@@ -249,11 +246,11 @@
 	display: flex;
 	min-width: 0;
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		flex-basis: 50%;
 		margin-bottom: rem( 16px );
 	}
-	@include breakpoint( "<660px" ) {
+	@include breakpoint( '<660px' ) {
 		margin-bottom: rem( 12px );
 	}
 
@@ -276,7 +273,7 @@
 			a.dops-notice__action {
 				margin-left: 0;
 				padding-left: 0;
-				@include breakpoint( "<480px" ) {
+				@include breakpoint( '<480px' ) {
 					text-transform: none;
 				}
 			}
@@ -294,11 +291,11 @@
 .jp-at-a-glance__left {
 	display: flex;
 
-	@include breakpoint( ">660px" ) {
+	@include breakpoint( '>660px' ) {
 		margin-right: rem( 16px );
 
 		&:last-child {
-			flex-basis: calc( 50% - .5rem ); // rem function doesn't work in calc()
+			flex-basis: calc( 50% - 0.5rem ); // rem function doesn't work in calc()
 		}
 	}
 }
@@ -333,7 +330,7 @@
 }
 .jp-connection-settings__info {
 	.gridicon {
-		opacity: .6;
+		opacity: 0.6;
 	}
 	.gridicon,
 	.jp-connection-settings__site-icon {
@@ -362,7 +359,7 @@
 }
 
 .jp-connection-settings__modal.dops-modal {
-	max-width: 450px;
+	max-width: 635px;
 }
 
 .jp-connection-settings__modal-body {
