+++ _inc/client/components/jetpack-termination-dialog/style.scss
@@ -1,8 +1,8 @@
 @import '../../scss/mixin_breakpoint.scss';
 @import '../../scss/variables/_colors.scss';
 
-.jetpack-disconnect-dialog__features,
-.jetpack-disconnect-dialog__survey  {
+.jetpack-termination-dialog__features,
+.jetpack-termination-dialog__survey  {
 	.dops-card {
 		margin: 0;
 	}
@@ -14,7 +14,7 @@
 
 }
 
-.jetpack-disconnect-dialog__header {
+.jetpack-termination-dialog__header {
 	h1 {
 		margin: 0;
 		font-size: 28px;
@@ -27,14 +27,14 @@
 
 }
 
-.jetpack-disconnect-dialog__info {
+.jetpack-termination-dialog__info {
 	font-size: 16px;
 	line-height: 24px;
 	// dops-card padding is enough
 	margin-top: 0;
 }
 
-.jetpack-disconnect-dialog__features-list {
+.jetpack-termination-dialog__features-list {
 	position: relative;
 	left: calc( -1em );
 	display: flex;
@@ -43,7 +43,7 @@
 	width: calc( 100% + 2em );
 }
 
-.jetpack-disconnect-dialog__feature {
+.jetpack-termination-dialog__feature {
 	background: $white;
 	border: 1px solid rgba( 213, 213, 213, 0.5 );
 	box-sizing: border-box;
@@ -55,7 +55,7 @@
 	}
 }
 
-.jetpack-disconnect-dialog__feature-header {
+.jetpack-termination-dialog__feature-header {
 	text-transform: uppercase;
 	font-size: 12px;
 	line-height: 12px;
@@ -72,10 +72,10 @@
 	}
 }
 
-.jetpack-disconnect-dialog__feature-body {
+.jetpack-termination-dialog__feature-body {
 	padding: 0 13px 0 13px;
 
-	.jetpack-disconnect-dialog__feature-body-amount {
+	.jetpack-termination-dialog__feature-body-amount {
 		font-weight: bold;
 		font-size: 24px;
 		line-height: 26px;
@@ -84,7 +84,7 @@
 	}
 }
 
-.jetpack-disconnect-dialog__get-help {
+.jetpack-termination-dialog__get-help {
 	font-style: normal;
 	font-weight: normal;
 	font-size: 16px;
@@ -92,7 +92,7 @@
 	text-align: center;
 }
 
-.jetpack-disconnect-dialog__button-row {
+.jetpack-termination-dialog__button-row {
 	display: flex;
 	flex-direction: row;
 	justify-content: space-between;
@@ -102,6 +102,6 @@
 	}
 }
 
-.jetpack-disconnect-dialog__close-icon {
+.jetpack-termination-dialog__close-icon {
 	color: #C8C8C8;
 }
