+++ _inc/client/rest-api/index.js
@@ -219,6 +219,12 @@ function JetpackRestApiClient( root, nonce ) {
 				.then( parseJsonResponse )
 				.then( body => JSON.parse( body.data ) ),
 
+		fetchSiteBenefits: () =>
+			getRequest( `${ apiRoot }jetpack/v4/site/benefits`, getParams )
+				.then( checkStatus )
+				.then( parseJsonResponse )
+				.then( body => JSON.parse( body.data ) ),
+
 		fetchRewindStatus: () =>
 			getRequest( `${ apiRoot }jetpack/v4/rewind`, getParams )
 				.then( checkStatus )
